deploy/requirements.txt
streamlit
tensorflow==2.10.0 # Pinning to a stable version
numpy
Pillow